@@ -15,34 +16,33 @@
     "@mui/material": "^6.1.10",
     "@mui/x-charts": "^7.23.1",
     "@mui/x-date-pickers": "^7.23.1",
-    "@reduxjs/toolkit": "^2.0.1",
-    "@storybook/addon-actions": "^8.1.1",
-    "@storybook/addon-themes": "^7.6.17",
-    "formik": "^2.4.5",
-    "i18next": "^23.7.16",
-    "i18next-browser-languagedetector": "^7.2.0",
-    "i18next-http-backend": "^2.4.2",
-    "luxon": "^3.4.4",
+    "@reduxjs/toolkit": "^2.4.0",
+    "@storybook/addon-actions": "^8.4.7",
+    "@storybook/addon-themes": "^8.4.7",
+    "formik": "^2.4.6",
+    "i18next": "^24.0.5",
+    "i18next-browser-languagedetector": "^8.0.2",
+    "i18next-http-backend": "^3.0.1",
+    "luxon": "^3.5.0",
     "notistack": "^3.0.1",
     "prop-types": "^15.8.1",
-    "react": "^18.2.0",
+    "react": "^19.0.0",
     "react-avatar-editor": "^13.0.2",
-    "react-dom": "^18.2.0",
-    "react-dropzone": "^14.2.3",
-    "react-i18next": "^14.0.0",
-    "react-redux": "^9.1.0",
+    "react-dom": "^19.0.0",
+    "react-dropzone": "^14.3.5",
+    "react-i18next": "^15.1.4",
+    "react-redux": "^9.1.2",
     "react-router": "^7.0.2",
-    "react-scripts": "5.0.1",
-    "recharts": "^2.12.7",
+    "recharts": "^2.14.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "swiper": "^11.1.15",
-    "uuid": "^9.0.1",
-    "yup": "^1.3.3"
+    "uuid": "^11.0.3",
+    "yup": "^1.5.0",
+    "swiper": "^11.1.15"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
+    "dev": "vite",
+    "build": "vite build",
     "server-build": "node build-script.js",
     "serve-build": "node ./server-build/server.js",
     "test": "react-app-rewired test",
@@ -87,40 +87,43 @@
     ]
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@storybook/addon-essentials": "^7.6.17",
-    "@storybook/addon-interactions": "^7.6.17",
-    "@storybook/addon-links": "^7.6.17",
-    "@storybook/addon-onboarding": "^1.0.11",
-    "@storybook/blocks": "^7.6.17",
-    "@storybook/preset-create-react-app": "^7.6.17",
-    "@storybook/react": "^7.6.17",
-    "@storybook/react-webpack5": "^7.6.17",
-    "@storybook/test": "^7.6.17",
+    "@storybook/addon-essentials": "^8.4.7",
+    "@storybook/addon-interactions": "^8.4.7",
+    "@storybook/addon-links": "^8.4.7",
+    "@storybook/addon-onboarding": "^8.4.7",
+    "@storybook/blocks": "^8.4.7",
+    "@storybook/preset-create-react-app": "^4.1.0",
+    "@storybook/react": "^8.4.7",
+    "@storybook/react-webpack5": "^8.4.7",
+    "@storybook/test": "^8.4.7",
+    "@vitejs/plugin-react": "^4.3.4",
     "css-loader": "^7.1.2",
-    "eslint": "^8.57.1",
+    "eslint": "^9.16.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.37.2",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-storybook": "^0.8.0",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "eslint-plugin-storybook": "^0.11.1",
+    "globals": "^15.12.0",
     "husky": "^9.1.7",
-    "jsdoc": "^4.0.2",
-    "lint-staged": "^15.2.10",
+    "jsdoc": "^4.0.4",
+    "lint-staged": "^15.2.11",
     "postcss": "^8.4.49",
     "postcss-scss": "^4.0.9",
-    "prettier": "^3.4.1",
+    "prettier": "^3.4.2",
     "react-app-rewired": "^2.2.1",
-    "sass": "^1.80.6",
-    "storybook": "^7.6.17",
+    "sass": "^1.82.0",
+    "storybook": "^8.4.7",
     "style-loader": "^4.0.0",
     "stylelint": "^16.11.0",
-    "stylelint-config-standard": "^36.0.0",
-    "stylelint-config-standard-scss": "^13.0.0",
+    "stylelint-config-standard": "^36.0.1",
+    "stylelint-config-standard-scss": "^14.0.0",
     "stylelint-scss": "^6.10.0",
-    "webpack": "^5.90.3"
+    "vite": "^6.0.3",
+    "vite-plugin-svgr": "^4.3.0"
   }
 }
