@@ -1,7 +1,5 @@
-import React from 'react';
-
 import { RouterProvider } from 'react-router/dom';
-import router from './router/router';
+import router from './router/router.jsx';
 
 export function App() {
   return <RouterProvider router={router} />;
