@@ -1,4 +1,4 @@
-import React, { useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Box, IconButton, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
@@ -76,6 +76,7 @@ const WorkExperienceModal = () => {
       setResponsibilities([]);
       resetForm();
       handleClose();
+      // eslint-disable-next-line no-unused-vars
     } catch (error) {
       const errorKey = modalData?.id
         ? 'modalNotifyText.workExperience.edit.error'
