@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Badge, IconButton, Popover } from '@mui/material';
 import { useSelector } from 'react-redux';
 import { useGetNotificationsQuery } from '../../../redux/services/notificationsApiSlice';
-import { ReactComponent as BellNotification } from '../../../assets/icons/bell.svg';
+import BellNotification from '../../../assets/icons/bell.svg?react';
 import { selectCurrentUser } from '../../../redux/auth/authSlice';
 import emptyNotificationLight from '../../../utils/constants/notification/whiteThemeIcons';
 import emptyNotificationDark from '../../../utils/constants/notification/darkThemeIcons';
