@@ -1,4 +1,4 @@
-import React, { useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { Box, IconButton, Link, Typography } from '@mui/material';
 import PropTypes from 'prop-types';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
@@ -57,6 +57,7 @@ const EducationItem = ({ id, type, name, description, startYear, endYear, icon: 
           horizontal: 'right',
         },
       });
+      // eslint-disable-next-line no-unused-vars
     } catch (error) {
       enqueueSnackbar(t('modalNotifyText.achievement.delete.error'), {
         variant: 'error',
