@@ -2,7 +2,7 @@ import AddIcon from '@mui/icons-material/Add';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import StarIcon from '@mui/icons-material/Star';
 import { Box, IconButton, Typography } from '@mui/material';
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { useSnackbar } from 'notistack';
@@ -99,6 +99,7 @@ const SpecializationCategories = () => {
       enqueueSnackbar(t('modalNotifyText.specialization.change.success', { name: activeSpecialization.name }), {
         variant: 'success',
       });
+      // eslint-disable-next-line no-unused-vars
     } catch (error) {
       enqueueSnackbar(t('modalNotifyText.specialization.change.error'), { variant: 'error' });
     }
