@@ -1,6 +1,6 @@
 import { Box, Typography } from '@mui/material';
 import ButtonGroup from '@mui/material/ButtonGroup';
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { setActiveMastery } from '../../../../redux/specialization/activeMasterySlice';
