@@ -6,7 +6,7 @@ import { clearTokens, setTokens } from '../../auth/tokenSlice';
 import { getTokenInHeaders } from '../../../utils/helpers';
 
 const baseQuery = fetchBaseQuery({
-  baseUrl: process.env.REACT_APP_API_URL,
+  baseUrl: import.meta.env.VITE_API_URL,
   prepareHeaders: (headers, { getState, endpoint }) => {
     if (PUBLIC_ENDPOINTS_ARRAY.includes(endpoint)) return;
 
@@ -52,7 +52,10 @@ const baseQueryWithReauth = async (args, api, extraOptions) => {
           }
         }
       }
+      // eslint-disable-next-line no-unused-vars
     } catch (error) {
+      api.dispatch(logOut());
+      api.dispatch(clearTokens());
       return result;
     }
   }
