@@ -1,10 +1,10 @@
 import { Link } from '@mui/icons-material';
-import { ReactComponent as Telegram } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/telegram.svg';
-import { ReactComponent as Linkedin } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/linkedin.svg';
-import { ReactComponent as Github } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/github.svg';
-import { ReactComponent as Behance } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/behance.svg';
-import { ReactComponent as Email } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/mail.svg';
-import { ReactComponent as Phone } from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/phone.svg';
+import Telegram from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/telegram.svg?react';
+import Linkedin from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/linkedin.svg?react';
+import Github from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/github.svg?react';
+import Behance from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/behance.svg?react';
+import Email from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/mail.svg?react';
+import Phone from '../../../assets/icons/ProfileRightSection/WhiteThemeIcons/phone.svg?react';
 
 export default {
   TELEGRAM_LINK: Telegram,
